043: break up text classes
