+++ src/components/display/text_field.js
@@ -1,0 +1,156 @@
+import Phaser from "phaser"
+import _ from "lodash"
+import TextDisplay from "./text_display"
+/**
+ * Draws an interactive button in the display
+ */
+class TextField extends TextDisplay {
+  /**
+   * @param {object} scene - The container Phaser.Scene
+   * @param {number} x - The x position of the TextDisplay in the game world
+   * @param {number} y - The y position of the TextDisplay in the game world
+   * @param {number} [width=200] - Display width in pixels
+   * @param {number} [height=400] - Display height in pixels
+   * @param {string} [content=""] - The text displayed
+   */
+  constructor(
+    scene,
+    x, y,
+    width=400,
+    height=200,
+    initialText="",
+    fontSize=24,
+    fontFamily="Helvetica",
+    outline=true
+  ) {
+
+    // group attributes
+    super(scene, x, y, width, height, "", fontSize, fontFamily, outline)
+
+    // add cursor
+    this.cursor = this.scene.add.text(x - this.width / 2, y - this.height / 2, "_", this.defaultStyles)
+    this.cursor.setOrigin(0, 0)
+    this.cursor.setMask(this.contentMask)
+
+    this.scene.add.tween({
+      targets: [this.cursor],
+      alpha: 0,
+      duration: 250,
+      yoyo: true,
+      repeat: -1
+    })
+
+    // set custom word wrapping to account for cursor
+    this.content.setWordWrapCallback(
+      function(text) {
+        let wrappedText = this.content.basicWordWrap(
+          (this.cursor ? text + "_" : text),
+          this.content.context,
+          this.width - this.content.padding.right * 2
+        ).trimRight().slice(0, -1)
+        return wrappedText
+      }, this
+    )
+
+    // calc pixel width of space char, for cursor
+    let testNoSpace = this.scene.add.text(0, 0, "II", this.defaultStyles)
+    let testSpace = this.scene.add.text(0, 0, "I I", this.defaultStyles)
+    this.spacePixelWidth = testSpace.width - testNoSpace.width
+    testNoSpace.destroy()
+    testSpace.destroy()
+
+    // set up listener for button press event
+    this.scene.events.on("domlessButtonPress", function(buttonChar, keyCode) {
+      this.addText(buttonChar, keyCode)
+    }, this)
+
+    this.addText(initialText)
+    
+  }
+
+  placeCursor() {
+    let wrappedText = this.content.basicWordWrap(
+      this.content.text + "_",
+      this.content.context,
+      this.width - this.content.padding.right * 2).split("\n")
+
+    let lastLineContent = _.last(wrappedText).trimRight().slice(0, -1)
+    let lastLine = this.scene.add.text(0, 0, lastLineContent, this.defaultStyles)
+    let lastLineCleaned = lastLineContent.replace(/(^ +| +$)/g, "").replace(/ {2,}\b/g, " ")
+    let lastLineSpaces = lastLineContent.length - lastLineCleaned.length
+    let lastLineWidth = lastLine.width + lastLineSpaces * this.spacePixelWidth
+    this.cursor.x = this.content.x + lastLineWidth - this.content.padding.right * 2
+    lastLine.destroy()
+
+    // calc cursor y pos
+    let contentHeight = this.content.height
+    if (this.content.height > this.height) {
+      contentHeight = this.height
+    }
+    this.cursor.y = this.y - this.height / 2 + contentHeight - this.cursor.height
+
+    // modify cursor position if we overflowed a line
+    if (this.cursor.x + this.cursor.width > this.x + this.width / 2) {
+      this.cursor.x = this.content.x - this.content.width / 2
+      this.cursor.y = this.y - this.height / 2 + contentHeight
+    }
+
+  }
+
+  addText(extraText, keyCode) {
+
+    this.pageDownButton.disableInput(true)
+
+    if (keyCode === Phaser.Input.Keyboard.KeyCodes.BACKSPACE) {
+      this.content.setText(this.content.text.slice(0, -1))
+    } else if (extraText) {
+      this.content.setText(this.content.text + extraText)
+    }
+    this.content.updateText()
+    if (this.content.height > this.height) {
+      this.content.y = this.y - (this.height / 2 + (this.content.height - this.height))
+    } else {
+      this.content.y = this.y - this.height / 2
+    }
+    this.placeCursor()
+
+    // enable the pageUp button if we have overflow content
+    // and it is not already enabled
+    if (this.content.y < this.y - this.height / 2 && !this.pageUpButton.alpha) {
+      this.pageUpButton.enableInput(true)
+    }
+  }
+
+  pageUp() {
+    // assume we are scrolling up a page
+    let scrollY = this.height
+
+    // modify behavior if we are near the top of the content
+    let disablePageUp = false
+    if (this.content.y >= this.y - this.height - this.height / 2) {
+      scrollY = this.y - this.content.y - this.height / 2
+      disablePageUp = true
+    }
+
+    // tween it by calling the parent method
+    super.pageUp(scrollY, null, disablePageUp)
+  }
+
+  pageDown() {
+    // assume we a scrolling down a page
+    let contentBottom = this.content.y + this.content.height
+    let scrollY = this.height
+
+    // modify behavior if we are near the bottom of the content
+    let disablePageDown = false
+    if (contentBottom <= this.y + this.height + this.height / 2) {
+      scrollY = contentBottom - this.y - this.height / 2
+      disablePageDown = true
+    }
+
+    // tween it by calling the parent method
+    super.pageDown(scrollY, null, disablePageDown)
+  }
+}
+
+export default TextField
